assignment3/requirements.txt
aiohttp
faker
python-dotenv